# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Ticket API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/users/GetUserByID/": {
            "post": {
                "description": "Returns the requested user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "description": "User ID",
                        "name": "userID",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.IDRequestInt64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errx.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errx.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errx.APIError"
                        }
                    }
                }
            }
        },
        "/users/GetUserByUsername/": {
            "post": {
                "description": "Returns the requested user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user by username",
                "parameters": [
                    {
                        "description": "Username",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UsernameDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errx.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errx.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errx.APIError"
                        }
                    }
                }
            }
        },
        "/users/GetUsersByIDs/": {
            "post": {
                "description": "Returns a list of all users requested",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get all users by IDs",
                "parameters": [
                    {
                        "description": "Users IDs",
                        "name": "usersIDs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserIDsDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.UserDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errx.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errx.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errx.APIError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.IDRequestInt64": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UserDTO": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "departmentId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.UserIDsDTO": {
            "type": "object",
            "required": [
                "ids"
            ],
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "dto.UsernameDTO": {
            "type": "object",
            "required": [
                "username"
            ],
            "properties": {
                "username": {
                    "type": "string"
                }
            }
        },
...
        "errx.APIError": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/errx.Error"
                }
            }
        },
        "errx.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/errx.ErrorCode"
                },
                "debug": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "stack": {
                    "type": "string"
                }
            }
        },
        "errx.ErrorCode": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12,
                13,
                14,
                15,
                16,
                17,
                18,
                19,
                20,
                21,
                22,
                23,
                24,
                25,
                26,
                27,
                28,
                29
            ],
            "x-enum-varnames": [
                "ErrInternalServerError",
                "ErrTicketNotFound",
                "ErrUnauthorized",
                "ErrInvalidInput",
                "ErrDuplicate",
                "ErrBadRequest",
                "ErrUserNotFound",
                "ErrTicketTypeNotFound",
                "ErrDepartmentNotFound",
                "ErrUserDuplicate",
                "ErrInvalidCredentials",
                "ErrWeakJWTSecret",
                "ErrIncorrectCaptcha",
                "ErrExpiredCaptcha",
                "ErrTicketStatusNotFound",
                "ErrTooManyRequest",
                "ErrApiKeyNotFound",
                "ErrServiceUnavailable",
                "ErrFileNotFound",
                "ErrLinkExpired",
                "ErrUnsupportedFileExtension",
                "ErrMaxFileSizeExceeded",
                "ErrMaxTicketFilesExceeded",
                "ErrRequestBodyTooLarge",
                "ErrForbidden",
                "ErrNotFound",
                "ErrOTPExpired",
                "ErrOTPInvalid",
                "ErrOTPSendFailed",
                "ErrOTPMaxAttemptsExceeded"
            ]
        }
    }
}